// servidor-local/src/utils/types.d.ts
export declare enum Role {
    CLIENTE = "cliente",
    ADMIN = "admin",
    PRESTADOR = "prestador",
    ADMIN = "administrador",
    EMPRESA = "empresa"
}